Build Rust via llc
Rather than having rustc emit object files, have it emit LLVM IR, and
compile that with llc.  The LLVM IR has no references to the problematic
__morestack.

Use sed to remove and replace bad parts of the IR.

The first, the dereferenceable() paramater, is added in a very new
version of LLVM, newer than my llc, so it must be removed.

The second is a static constant from libcore's str module.  rustc
generates IR stating that it is "available externally", but it also
includes the value.  Since the value is available, simply switch it it
"internal".