Switch back to main test program and run!
Remove the simple iterator test, switching back to the main VGA test
program.  It actually successfully runs and prints!

The .bss section for Rust is not zeroed, so there is currently a helper
function in the vga module to fix up some globals.

The iterator in write() does not work, it simply returns without
iterating.  This may be due to the uncleared .bss.  For now, putc is
used directly.